fetch/metadata/generated/worker-dedicated-importscripts-sec-fetch-dest.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <script type="module">
  'use strict';
  function induceRequest(path, options) {
    const src = `
      importScripts('${path}&mime=text/javascript');
      postMessage('done');
    `;
    const url = URL.createObjectURL(new Blob([src], { type: 'application/javascript' }));
    return new Promise((resolve, reject) => {
      const worker = new Worker(url, options);
      worker.onmessage = resolve;
      worker.onerror = reject;
    });
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest('https://{{host}}:{{ports[https][0]}}' + path)
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'script');
        });
  }, 'sec-fetch-dest - ');
  </script>
</html>
